refactor: share one return path check across the routes that take one
localReturnPath moves to utils and returnUrlSchema to models/common, so
sign in and cookie preferences validate the same way. The query parameter
is returnUrl everywhere, matching the name cookie preferences already used.

Cookie preferences previously used isPathRelative, which is startsWith('/').
That let //host and /.//host through to h.redirect, both of which a browser
follows off site. Empty is now rejected too; the layout builds the value
from currentPath, which always has a path.
