+++ src/server/plugins/router.ts
@@ -1,7 +1,6 @@
 import {
   getCacheService,
-  handleLegacyRedirect,
-  isPathRelative
+  handleLegacyRedirect
 } from '@defra/forms-engine-plugin/engine/helpers.js'
 import {
   crumbSchema,
@@ -30,6 +29,7 @@ import { type CookieConsent } from '~/src/common/types.js'
 import { config } from '~/src/config/index.js'
 import { FORM_PREFIX } from '~/src/server/constants.js'
 import { t as runnerT } from '~/src/server/i18n/index.js'
+import { returnUrlSchema } from '~/src/server/models/common.js'
 import { getErrorPreviewHandler } from '~/src/server/plugins/error-preview/error-preview.js'
 import {
   authRoutes,
@@ -264,10 +264,6 @@ export default {
           const { slug } = params
           let { returnUrl } = query
 
-          if (returnUrl && !isPathRelative(returnUrl)) {
-            throw Boom.badRequest('Return URL must be relative')
-          }
-
           const analyticsDecision = (
             payload['cookies[analytics]'] ?? ''
           ).toLowerCase()
@@ -313,7 +309,7 @@ export default {
               'cookies[dismissed]': Joi.string().valid('yes', 'no').optional()
             }),
             query: Joi.object({
-              returnUrl: Joi.string().optional()
+              returnUrl: returnUrlSchema.optional()
             })
           }
         }
