+++ src/server/utils/utils.test.js
@@ -4,6 +4,7 @@ import { config } from '~/src/config/index.js'
 import {
   applyTraceHeaders,
   getFeedbackFormLink,
+  localReturnPath,
   resolveLanguage
 } from '~/src/server/utils/utils.js'
 
@@ -101,5 +102,90 @@ describe('utils', () => {
         resolveLanguage({}, { id: 'session-id', set: yarSet, get: yarGet })
       ).toBe('cy')
     })
+  })
+})
+
+describe('localReturnPath', () => {
+  it('resolves a local path to itself', () => {
+    expect(localReturnPath('/homepage/test-form')).toBe('/homepage/test-form')
+    expect(localReturnPath('/form/test-form/page-one?lang=cy')).toBe(
+      '/form/test-form/page-one?lang=cy'
+    )
+  })
+
+  it('returns the parser’s normalised form of a path carrying a control character, not the raw input', () => {
+    // Node rejects a Location header containing a newline. The parser drops
+    // it, so the resolved path is safe to send.
+    expect(localReturnPath('/foo\nSet-Cookie: a=b')).toBe(
+      '/fooSet-Cookie:%20a=b'
+    )
+  })
+
+  it('has no resolved path for anything that could leave the service', () => {
+    expect(localReturnPath('//evil.example')).toBeNull()
+    expect(localReturnPath('https://evil.example')).toBeNull()
+    expect(localReturnPath('homepage/test-form')).toBeNull()
+    expect(localReturnPath('')).toBeNull()
+    expect(localReturnPath(undefined)).toBeNull()
+    // A URL parser treats a leading backslash as a slash, so this resolves
+    // to protocol-relative `//evil.example` even though the raw string
+    // starts with a single `/`.
+    expect(localReturnPath('/\\evil.example')).toBeNull()
+    // A URL parser strips ASCII tab and newline anywhere in the input before
+    // resolving, so these also collapse to `//evil.example`.
+    expect(localReturnPath('/\t/evil.example')).toBeNull()
+    expect(localReturnPath('/\n/evil.example')).toBeNull()
+  })
+
+  it.each([
+    '/.//evil.example',
+    '/..//evil.example',
+    '/./\\evil.example',
+    '/../\\evil.example',
+    '/.///evil.example',
+    '/x/..//evil.example'
+  ])(
+    'has no resolved path for %s, whose dot segment keeps the local origin while resolving to a protocol-relative path',
+    (payload) => {
+      // The dot segment resolves inside the origin, so the origin does not
+      // change and the input has one leading slash. It resolves to the path
+      // `//evil.example`, which a browser reads as a host.
+      expect(localReturnPath(payload)).toBeNull()
+    }
+  )
+
+  it('lets no separator and prefix combination resolve to an off-site path', () => {
+    const payloads = []
+    const seps = ['/', '\\', '\t', '\n', '\r', '%2F', '%5C', ' ']
+    const prefixes = ['', '/', '//', '///', '/.', '/..', '/./', '/../', '/x/..']
+
+    for (const prefix of prefixes) {
+      for (const sep of seps) {
+        for (const host of ['attacker.example', 'evil.example']) {
+          payloads.push(
+            `${prefix}${sep}${host}`,
+            `${prefix}${sep}${sep}${host}`,
+            `${prefix}${sep}${sep}${sep}${host}`
+          )
+        }
+      }
+    }
+
+    // An accepted value must be a path: one leading slash and no scheme.
+    const leaks = payloads
+      .map((payload) => [payload, localReturnPath(payload)])
+      .filter(
+        ([, path]) =>
+          path !== null &&
+          (path.startsWith('//') || /^[a-z][a-z0-9+.-]*:/i.test(path))
+      )
+
+    expect(leaks).toEqual([])
+  })
+
+  it('takes a path, so a fully qualified URL is not one', () => {
+    expect(
+      localReturnPath('http://localhost:3009/homepage/test-form')
+    ).toBeNull()
   })
 })
